Tests adapted for new workflow
